RF:  make `data.stats` module
