+++ eelbrain/data/stats/__init__.py
@@ -1,0 +1,1 @@
+from .stats import *
