+++ eelbrain/data/stats/test.py
@@ -4,7 +4,7 @@ statistical tests for data objects
 """
 from __future__ import division
 
-import random, itertools
+import itertools
 
 import numpy as np
 import scipy.stats
@@ -14,6 +14,7 @@ from ... import fmtxt
 from ..data_obj import (isvar, isndvar, asvar, isfactor, asfactor, assub,
                         ismodel, iscategorial, ascategorial, cellname,
                         Celltable)
+from .permutation import resample
 
 
 __test__ = False
@@ -800,53 +801,3 @@ class bootstrap_pairwise(object):
         test = self.t_resampled[:, None] > np.abs(t)
         p = np.sum(test, axis=0) / self.t_resampled.shape[0]
         return p
-
-
-
-def resample(Y, samples=10000, replacement=False, unit=None):
-    """
-    Generator function to resample a dependent variable (Y) multiple times
-
-    Parameters
-    ----------
-    Y : Var | NDVar
-        Variable which is to be resampled.
-    samples : int
-        number of samples to yield.
-    replacement : bool
-        whether random samples should be drawn with replacement or without.
-    unit : categorial
-        Factor specifying unit of measurement (e.g. subject). If unit is
-        specified, resampling proceeds by first resampling the categories of
-        unit (with or without replacement) and then shuffling the values
-        within units (no replacement).
-
-    Returns
-    -------
-    Iterator over Y_resampled. The same copy of ``Y`` is yielded in each
-    iteration with different data.
-    """
-    samples = int(samples)
-    if isvar(Y):
-        pass
-    elif isndvar(Y):
-        if not Y.has_case:
-            raise ValueError("Need NDVar with cases")
-    else:
-        raise TypeError("need Var or NDVar")
-
-    Yout = Y.copy('{name}_resampled')
-
-    if unit:
-        raise NotImplementedError("Check implementation")
-    else:
-        if replacement:
-            N = len(Y)
-            for _ in xrange(samples):
-                index = np.random.randint(N, N)
-                Yout.x = Y.x[index]
-                yield Yout
-        else:  # OK
-            for _ in xrange(samples):
-                np.random.shuffle(Yout.x)
-                yield Yout
