--- eelbrain/data/test/__init__.py
@@ -1,5 +1,0 @@
-__test__ = False
-
-from .glm import anova, ancova
-from .test import (pairwise, ttests, oneway, correlations, bootstrap_pairwise,
-                   lilliefors)
